Try to improve MiniMax algorithm and implement tests
PlayerDev: added quiesence search to continue until a leaf is a quiet position in order to reduce misjudging a position in the evaluation.
Evaluation: Added evaluator based on Piece Square Tables with some crude first version of values for early and late game.
Computer Play: Added new mode of computer 1 versus computer 2 with result and game logging.
Added ray utils to let computer play in parallel on a few cores, added some positions and game pgns for this purpose.